requirements.txt
grpcio>=1.31.0
protobuf>=3.13.0
google-api==0.1.12
google-api-core==1.22.1
ecdsa>=0.16.1
